ai_services/food_classification_service_mock.py
import os
import json
import random
from flask import Flask, request, jsonify
import base64
import io
from PIL import Image

app = Flask(__name__)

# Constants
NUTRIENTS_DB_PATH = 'food_nutrients_db.json'

# Load nutrients database
def load_nutrients_db():
    if os.path.exists(NUTRIENTS_DB_PATH):
        with open(NUTRIENTS_DB_PATH, 'r') as f:
            return json.load(f)
    else:
        # Default empty database
        return {}

# Function to generate a mock classification response
def mock_classification_response():
    print("Generating mock classification response")
    # Use the nutrients database to get a random food item
    nutrients_db = load_nutrients_db()
    if not nutrients_db:
        return jsonify({'error': 'Nutrients database not available.'}), 500
        
    # Get a random food item from the nutrients database
    food_items = list(nutrients_db.keys())
    if not food_items:
        return jsonify({'error': 'No food items in the nutrients database.'}), 500
        
    predicted_class = random.choice(food_items)
    confidence = 0.85  # Mock confidence
    
    # Get nutrients information
    nutrients_info = nutrients_db.get(predicted_class, {})
    
    # Prepare response
    response = {
        'food': predicted_class,
        'confidence': confidence,
        'calories': nutrients_info.get('calories', 0),
        'nutrients': nutrients_info.get('nutrients', {}),
        'description': nutrients_info.get('description', f'This appears to be {predicted_class}')
    }
    
    return jsonify(response)

# API endpoint for image classification
@app.route('/classify', methods=['POST'])
def classify_image():
    if 'image' not in request.json:
        return jsonify({'error': 'No image provided'}), 400
    
    try:
        # Get the base64 encoded image
        img_data = request.json['image']
        print(f"Received image data of length: {len(img_data)}")
        
        # Handle potential padding issues with base64
        # Add padding if needed
        padding = 4 - (len(img_data) % 4) if len(img_data) % 4 != 0 else 0
        img_data += '=' * padding
        
        # Remove data URL prefix if present
        if img_data.startswith('data:image'):
            img_data = img_data.split(',')[1]
        
        try:
            img_bytes = base64.b64decode(img_data)
            print(f"Successfully decoded base64 data, length: {len(img_bytes)} bytes")
            
            try:
                img = Image.open(io.BytesIO(img_bytes))
                print(f"Successfully opened image: {img.format}, size: {img.size}, mode: {img.mode}")
            except Exception as e:
                print(f"Error opening image: {str(e)}")
                # Return a mock response instead of failing
                return mock_classification_response()
        except Exception as e:
            print(f"Error decoding base64: {str(e)}")
            # Return a mock response instead of failing
            return mock_classification_response()
        
        # Always use mock implementation
        return mock_classification_response()
        
    except Exception as e:
        print(f"Error in classify_image: {str(e)}")
        return jsonify({'error': str(e)}), 500

# API endpoint for creating a sample nutrients database
@app.route('/create_nutrients_db', methods=['POST'])
def create_nutrients_db():
    try:
        # Create a sample nutrients database
        nutrients_db = {
            "apple_pie": {
                "calories": 237,
                "nutrients": {
                    "carbohydrates": 33.6,
                    "protein": 2.4,
                    "fat": 11.0,
                    "fiber": 1.4,
                    "sugar": 18.9
                },
                "description": "Apple pie with a sweet filling of apple, sugar, and cinnamon"
            },
            "pizza": {
                "calories": 266,
                "nutrients": {
                    "carbohydrates": 33.0,
                    "protein": 11.0,
                    "fat": 10.0,
                    "fiber": 2.3,
                    "sugar": 3.6
                },
                "description": "Pizza with cheese, tomato sauce, and various toppings"
            },
            "salad": {
                "calories": 152,
                "nutrients": {
                    "carbohydrates": 11.0,
                    "protein": 3.8,
                    "fat": 11.0,
                    "fiber": 3.0,
                    "sugar": 2.5
                },
                "description": "Mixed greens with vegetables and dressing"
            },
            # Add more food items as needed
        }
        
        # Save the nutrients database
        with open(NUTRIENTS_DB_PATH, 'w') as f:
            json.dump(nutrients_db, f, indent=4)
        
        return jsonify({'message': 'Sample nutrients database created successfully'})
    except Exception as e:
        return jsonify({'error': str(e)}), 500

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5001, debug=True)