ai_services/requirements.txt
numpy
pandas
scikit-learn
flask
tensorflow
pillow
requests
matplotlib
